Fix an edge case that causes the salt to be lost
When synced profiles are encrypted for the 2nd time, the key is used
instead of the password, and no salt is saved. This causes the salt
value to be lost, which in turn causes new clients not to be able to
decrypt the synced data.

Ensure that the salt value is always stored.