Changed "usedText" to "usedURL"
In preparation for allowing the user to set an arbitrary “usedText”
string, the profile matched URL is now stored in the ‘alt’ attribute of
the input field.

This frees the usedText value so it can be used just for password
generation.

Other bits of code have been adjusted to use “usedURL” to avoid
confusion.