Fix blank password error
If I blank the password in the popup, then close and reopen, I get the
following error in console.

Error in event handler for 'undefined': Cannot read property 'length' of
undefined TypeError: Cannot read property 'length' of undefined
at byteArrayToString
(chrome-extension://nedpgficfnmjjgmgcphgfhlknmkgaoij/javascript/passwordmaker/aes.js:292:27)
at
chrome-extension://nedpgficfnmjjgmgcphgfhlknmkgaoij/javascript/settings.js:301:37

This stops the loading of the popup so it is missing the list of
profiles and other things.

I don't understand why it was passing this if statement as true, as the
array element is 'undefined', but this change has fixed it.