Condition invalid, fixed regression
Fixes password not saving on exit.

Well, really, its password not loading on startup. I checked the
password was being saving inthe SQLITE localstorage, that was fine.
However on load it was not meeting the condition to get the password :(

Don't know how I missed this, I must of reloaded a hundred times prior
to the pull request.