Adds Alt+Shift+M keyboard shortcut for invoking popup browser action
Using the Alt+Shift level here to not collide with default Chrome
key shortcuts that consistently use Ctrl/Command+Shift.
I chose M because it is easily reachable on the keyboard and because
it allows me to remember the shortcut by thinking of calling the
password[M]aker. Also, using Alt+Shift+P seemed a bit odd to me as
usually browser key commands including P invoke functions to do with
printing.