Added keyboard shortcut for Copy Password
This extends the keyboard handler to recognize Ctrl+C or Option+C
anywhere in the extension popup.  This copies the generated password
to the clipboard the same way the button `Copy Password` does.

Also moved the calls to updateFields, so that this will also be called when
the buttons are pressed.