Make handling lastError cross-platform
Chrome and Firefox both have chrome.runtime.lastError property, but they
are differently implemented.

On Firefox chrome.runtime.lastError can be "null" or instance of
"Error"[1].  However on Chrome it will be "string" with error message or
it will remain undefined[2].

Instead of checking whether chrome.runtime.lastError against undefined
we can check if it evaluates to true or false.

[1] https://developer.mozilla.org/en-US/Add-ons/WebExtensions/API/runtime/lastError
[2] https://developer.chrome.com/extensions/runtime#property-lastError